+++ examples/simple.rs
@@ -1,26 +1,17 @@
-use bevy::prelude::*;
-use bevy_button_released_plugin::{ButtonReleasedEvent, ButtonsReleasedPlugin};
+use bevy::{color::palettes::css::*, prelude::*};
+use bevy_button_released_plugin::{ButtonsReleasedPlugin, OnButtonReleased};
 
 pub fn main() {
-    let mut app = App::new();
-    app.add_plugins(DefaultPlugins)
+    App::new()
+        .add_plugins(DefaultPlugins)
         .add_plugins(ButtonsReleasedPlugin)
         .add_systems(Startup, setup)
-        .add_systems(Update, button_system);
-    app.run();
+        .run();
 }
 
-fn button_system(
-    mut reader: EventReader<ButtonReleasedEvent>,
-    q: Query<&Name>,
-    mut q_text: Query<&mut Text>,
-) {
-    let mut text = q_text.single_mut();
-    for event in reader.read() {
-        if let Ok(button_name) = q.get(**event) {
-            text.sections[0].value = format!("Last button released: {}", button_name);
-        }
-    }
+fn on_button(trigger: Trigger<OnButtonReleased>, q: Query<&Name>, mut q_text: Query<&mut Text>) {
+    let button_name = q.get(trigger.entity()).expect("Missing Name!");
+    q_text.single_mut().sections[0].value = format!("Last button released: {}", button_name);
 }
 
 fn setup(mut commands: Commands) {
@@ -28,47 +19,46 @@ fn setup(mut commands: Commands) {
     commands.spawn(Camera2dBundle::default());
 
     // root node
-    commands
-        .spawn(NodeBundle {
-            style: Style {
-                width: Val::Percent(100.0),
-                height: Val::Percent(100.0),
-                justify_content: JustifyContent::SpaceEvenly,
-                flex_direction: FlexDirection::Column,
-                ..default()
-            },
-            ..default()
-        })
-        .with_children(|parent| {
-            let style: Style = Style {
-                margin: UiRect::all(Val::Px(18.0)),
-                padding: UiRect::all(Val::Px(30.0)),
-                ..default()
-            };
-            parent.spawn(
-                TextBundle::from_section(
-                    "Press any button below",
-                    TextStyle {
-                        font_size: 30.0,
-                        ..default()
-                    },
-                )
-                .with_text_justify(JustifyText::Center)
-                .with_style(style.clone()),
-            );
-            for (text, color) in [
-                ("Green", Color::LinearRgba(LinearRgba { red: 0.0, green: 1.0, blue: 0.0, alpha: 1.0 })),
-                ("Red", Color::LinearRgba(LinearRgba{red: 1.0, green: 0.0, blue: 0.0, alpha: 1.0})),
-                ("Yellow", Color::LinearRgba(LinearRgba{red: 1.0, green: 1.0, blue: 0.0, alpha: 1.0})),
-            ] {
-                parent.spawn((
+    commands.spawn(root()).with_children(|parent| {
+        parent.spawn(top_text());
+        for (text, color) in [("Green", GREEN), ("Red", RED), ("Yellow", YELLOW)] {
+            parent
+                .spawn((
                     ButtonBundle {
-                        style: style.clone(),
-                        background_color: BackgroundColor(color),
+                        style: style(),
+                        background_color: BackgroundColor(color.into()),
                         ..default()
                     },
                     Name::new(text),
-                ));
-            }
-        });
+                ))
+                .observe(on_button);
+        }
+    });
+}
+fn top_text() -> TextBundle {
+    TextBundle::from_section("Press any button below", default())
+        .with_text_justify(JustifyText::Center)
+        .with_style(style())
+}
+
+fn root() -> NodeBundle {
+    NodeBundle {
+        style: Style {
+            width: Val::Percent(100.0),
+            height: Val::Percent(100.0),
+            justify_content: JustifyContent::SpaceEvenly,
+            flex_direction: FlexDirection::Column,
+            ..default()
+        },
+        ..default()
+    }
+}
+
+fn style() -> Style {
+    let rect = UiRect::all(Val::Px(30.0));
+    Style {
+        margin: rect,
+        padding: rect,
+        ..default()
+    }
 }