+++ src/lib.rs
@@ -1,18 +1,34 @@
+#![allow(deprecated)]
 use bevy::ecs::system::Res;
 #[cfg(feature = "auto_add")]
 use bevy::ecs::*;
 use bevy::input::touch::Touches;
 use bevy::reflect::Reflect;
 use bevy::{
-    prelude::{App, Changed, Component, Deref, Entity, Event, EventWriter, Plugin, Query, Update},
+    prelude::{App, Changed, Component, Entity, Event, Plugin, Query, Update},
     ui::*,
 };
+use system::Commands;
 
+/// Plugin that adds `OnButtonReleased` event for entities with `GameButton` and `Interaction`
+/// components when Interaction changes from `Interaction::Pressed`
+/// to `Interaction::Hovered`.
 pub struct ButtonsReleasedPlugin;
 
-#[derive(Deref, Event, Reflect)]
+#[deprecated(note = "ButtonReleasedEvent is deprecated, use `OnButtonReleased` instead.")]
+#[derive(bevy::prelude::Deref, Event, Reflect)]
 pub struct ButtonReleasedEvent(Entity);
 
+/// Event triggered when Interaction changes from `Interaction::Pressed`
+/// to `Interaction::Hovered` for entity that has GameButton component.
+/// It is using the observer pattern for events.
+#[derive(Event, Reflect)]
+pub struct OnButtonReleased;
+
+/// Component storing previous interaction state.
+/// When it is part of entity that also contains `Interaction` component
+/// it will allow triggering `OnButtonReleased` when Interaction changes from `Interaction::Pressed`
+/// to `Interaction::Hovered`.
 #[derive(Component, Default, Reflect)]
 pub struct GameButton {
     pub last_state: Interaction,
@@ -20,9 +36,9 @@ pub struct GameButton {
 
 impl Plugin for ButtonsReleasedPlugin {
     fn build(&self, app: &mut App) {
-        app.add_event::<ButtonReleasedEvent>()
-            .register_type::<GameButton>()
-            .register_type::<ButtonReleasedEvent>()
+        app.register_type::<GameButton>()
+            .add_event::<ButtonReleasedEvent>()
+            .register_type::<OnButtonReleased>()
             .add_systems(Update, button_click_system);
         #[cfg(feature = "auto_add")]
         app.add_systems(bevy::app::PostUpdate, add_game_button);
@@ -42,16 +58,17 @@ fn add_game_button(
 fn button_click_system(
     touches: Res<Touches>,
     mut interaction_query: Query<(Entity, &Interaction, &mut GameButton), Changed<Interaction>>,
-    mut ev: EventWriter<ButtonReleasedEvent>,
+    mut commands: Commands,
+    mut ev: event::EventWriter<ButtonReleasedEvent>,
 ) {
     let any_input_released = touches.any_just_released();
     for (entity, interaction, mut game_button) in &mut interaction_query {
         let was_hovered = any_input_released || *interaction == Interaction::Hovered;
         if was_hovered && game_button.last_state == Interaction::Pressed {
+            commands.trigger_targets(OnButtonReleased, entity);
             ev.send(ButtonReleasedEvent(entity));
         }
 
         game_button.last_state = *interaction;
     }
 }
-